server/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg==3.1.19
psycopg-binary==3.1.19
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
